+++ plugins/plotly-express/src/js/src/PlotlyExpressChartModel.ts
@@ -1,16 +1,4 @@
-import type {
-  Layout,
-  Data,
-  PlotData,
-  LayoutAxis,
-  PlotlyHTMLElement,
-  PlotMouseEvent,
-  PlotSelectionEvent,
-  LegendClickEvent,
-  ClickAnnotationEvent,
-  SunburstClickEvent,
-} from 'plotly.js';
-import Plotly from 'plotly.js-dist-min';
+import type { Layout, Data, PlotData, LayoutAxis } from 'plotly.js';
 import type { dh as DhType } from '@deephaven/jsapi-types';
 import {
   type DateTimeColumnFormatter,
@@ -53,80 +41,6 @@ import {
 } from './PlotlyExpressChartUtils';
 
 const log = Log.module('@deephaven/js-plugin-plotly-express.ChartModel');
-
-/**
- * The Plotly graph div element. Plotly's `PlotlyHTMLElement` type already models
- * `.on` (with per-event payload types) and `.data`; we add an overload for the
- * treemap/icicle click events, which share the sunburst click event shape but
- * aren't in the type despite existing at runtime.
- */
-type PlotlyGraphDiv = PlotlyHTMLElement & {
-  on: (
-    event:
-      | 'plotly_sunburstclick'
-      | 'plotly_treemapclick'
-      | 'plotly_icicleclick',
-    callback: (event: SunburstClickEvent) => void
-  ) => void;
-};
-
-/** Keyboard modifier keys held during a user interaction. */
-type EventModifiers = {
-  shift: boolean;
-  ctrl: boolean;
-  alt: boolean;
-  meta: boolean;
-};
-
-/**
- * Data-space field names that Plotly may include on point data at runtime,
- * depending on the chart type
- */
-const POINT_DATA_FIELDS = [
-  'x',
-  'y',
-  'z',
-  'lat',
-  'lon',
-  'location',
-  'r',
-  'theta',
-  'open',
-  'high',
-  'low',
-  'close',
-  'label',
-  'parent',
-  'value',
-  'id',
-  'text',
-] as const;
-
-/**
- * Serialize a Plotly point datum into the payload sent to Python.
- * Includes all data-space fields that are present on the point,
- * plus `trace_name`, `trace_type`, and `curve_number`.
- */
-function serializePoint(
-  p: Record<string, unknown> & { data: Partial<PlotData>; curveNumber: number }
-): Record<string, unknown> {
-  const result: Record<string, unknown> = {};
-  POINT_DATA_FIELDS.forEach(field => {
-    if (field in p && p[field] !== undefined) {
-      result[field] = p[field];
-    }
-  });
-  result.trace_name = p.data.name;
-  result.trace_type = p.data.type;
-  result.curve_number = p.curveNumber;
-  return result;
-}
-
-/** Hierarchical trace types handled by wireHierarchicalClickHandler. */
-const HIERARCHICAL_TYPES = new Set(['sunburst', 'treemap', 'icicle']);
-
-/** Plotly's default doubleClickDelay for legend click discrimination. */
-const LEGEND_DBL_CLICK_DELAY = 300;
 
 export class PlotlyExpressChartModel extends ChartModel {
   /**
@@ -378,11 +292,6 @@ export class PlotlyExpressChartModel extends ChartModel {
 
     this.isSubscribed = true;
 
-    // Track keyboard modifier keys from the raw DOM pointer event so they can
-    // be attached to every event payload (see getModifiers). Capture phase so
-    // it runs before Plotly dispatches its own handlers.
-    document.addEventListener('pointerdown', this.handlePointerModifiers, true);
-
     this.widgetUnsubscribe = this.widget.addEventListener<DhType.Widget>(
       this.dh.Widget.EVENT_MESSAGE,
       ({ detail }) => {
@@ -409,357 +318,6 @@ export class PlotlyExpressChartModel extends ChartModel {
       // there are filters, so the server expects the filter to be sent
       this.sendFilterUpdated(this.filterMap ?? new Map());
     }
-
-    // Wire up event callbacks. This is a no-op until the Chart component
-    // provides the Plotly graph div via setPlotElement().
-    this.wireEventCallbacks();
-  }
-
-  /**
-   * The Plotly graph div element, provided by the Chart component once Plotly
-   * has initialized (via its onInitialized callback).
-   */
-  private plotElement: PlotlyGraphDiv | null = null;
-
-  private eventListenersWired = false;
-
-  /**
-   * Keyboard modifier keys held during the most recent pointer interaction,
-   * captured from the raw DOM event (see subscribe). Attached to every event
-   * payload sent to Python so callbacks can branch on e.g. shift-click. We read
-   * these from the DOM directly rather than from Plotly's event objects, which
-   * don't expose modifier state for every event type (selection, relayout...).
-   */
-  private modifiers: EventModifiers = {
-    shift: false,
-    ctrl: false,
-    alt: false,
-    meta: false,
-  };
-
-  private handlePointerModifiers = (event: PointerEvent): void => {
-    this.modifiers = {
-      shift: event.shiftKey,
-      ctrl: event.ctrlKey,
-      alt: event.altKey,
-      meta: event.metaKey,
-    };
-  };
-
-  private getModifiers(): EventModifiers {
-    return { ...this.modifiers };
-  }
-
-  setPlotElement(plotElement: HTMLElement | null): void {
-    this.plotElement = plotElement as PlotlyGraphDiv | null;
-    this.eventListenersWired = false;
-    this.wireEventCallbacks();
-  }
-
-  /**
-   * Attach Plotly event listeners for the hierarchical drill-down clicks
-   * (sunburst/treemap/icicle). These must be wired imperatively rather than
-   * passed as props to the Plotly component because:
-   *   - Plotly only honors a handler's `return false` (to prevent the default
-   *     drill-down) when that handler is the last one registered, and
-   *   - react-plotly.js registers a `plotly_sunburstclick` "update" listener of
-   *     its own after onInitialized, which would otherwise shadow ours, and
-   *   - react-plotly.js exposes no treemap/icicle click props at all.
-   */
-  private wireEventCallbacks(): void {
-    // Pulling the plot element directly rather than passing handlers into the
-    // react component lets us guarantee ordering for the hierarchical clicks.
-    const gd = this.plotElement;
-    if (
-      gd == null ||
-      typeof gd.on !== 'function' ||
-      this.callbackMap.size === 0 ||
-      this.eventListenersWired
-    ) {
-      return;
-    }
-    this.eventListenersWired = true;
-
-    // Defer wiring to a microtask so our listener registers after
-    // react-plotly.js attaches its own, which clobbers our sunburst handler.
-    // Sunburst could be handled directly, but treemap and icicle are not
-    // supported by react-plotly.js at all so they are all handled here.
-    queueMicrotask(() => {
-      if (this.plotElement !== gd) {
-        // The element was swapped out or removed before the microtask ran.
-        return;
-      }
-      this.wireHierarchicalClickHandler(gd);
-    });
-  }
-
-  /**
-   * Wire the hierarchical (sunburst/treemap/icicle) click handler. Always
-   * prevents the default drill-down and re-triggers it via Plotly.animate only
-   * if the Python callback allows it.
-   */
-  private wireHierarchicalClickHandler(gd: PlotlyGraphDiv): void {
-    const clickId = this.callbackMap.get('on_click');
-    if (clickId == null || !this.isPreventable(clickId)) {
-      return;
-    }
-    const hierEvents = [
-      'plotly_sunburstclick',
-      'plotly_treemapclick',
-      'plotly_icicleclick',
-    ] as const;
-    hierEvents.forEach(eventName => {
-      gd.on(eventName, (data: SunburstClickEvent) => {
-        const args = {
-          points: data.points.map(p => ({
-            label: p.label,
-            parent: p.parent,
-            value: p.value,
-            id: p.id,
-            trace_name: (p.data as Partial<PlotData>).name,
-            trace_type: (p.data as Partial<PlotData>).type,
-            curve_number: p.curveNumber,
-          })),
-          next_level: data.nextLevel ?? null,
-          modifiers: this.getModifiers(),
-        };
-        this.sendEventCallbackWithResponse(clickId, args).then(allowed => {
-          if (allowed && data.nextLevel != null) {
-            const traceIdx = data.points?.[0]?.curveNumber ?? 0;
-            Plotly.animate(
-              gd,
-              { data: [{ level: data.nextLevel }], traces: [traceIdx] },
-              {
-                frame: { redraw: false, duration: 300 },
-                transition: { duration: 300, easing: 'cubic-in-out' },
-                mode: 'immediate',
-                fromcurrent: true,
-              }
-            );
-          }
-        });
-        return false; // Always prevent default drill-down
-      });
-    });
-  }
-
-  /**
-   * Handle a double-click on the plot area (e.g. reset axes in zoom/pan mode).
-   */
-  onDoubleClick(): void {
-    const doubleClickId = this.callbackMap.get('on_double_click');
-    if (doubleClickId == null) {
-      return;
-    }
-    this.sendEventCallback(doubleClickId, {
-      modifiers: this.getModifiers(),
-    });
-  }
-
-  /**
-   * Handle WebGL context lost.
-   */
-  onWebGlContextLost(): void {
-    const webGlLostId = this.callbackMap.get('on_web_gl_context_lost');
-    if (webGlLostId == null) {
-      return;
-    }
-    this.sendEventCallback(webGlLostId, {
-      modifiers: this.getModifiers(),
-    });
-  }
-
-  /**
-   * Handle a (non-hierarchical) point click.
-   * Hierarchical clicks are preventable and handled
-   * imperatively via wireHierarchicalClickHandler instead.
-   * Non-hierarchical traces in a layered/subplot figure that also
-   * contains hierarchical traces still arrive here, so we only skip
-   * when every clicked point is a hierarchical type.
-   */
-  onClick(event: PlotMouseEvent): void {
-    const clickId = this.callbackMap.get('on_click');
-    if (clickId == null) {
-      return;
-    }
-    if (this.isPreventable(clickId)) {
-      const allHierarchical = event.points.every(p =>
-        HIERARCHICAL_TYPES.has(p.data.type ?? '')
-      );
-      if (allHierarchical) {
-        return; // Handled by wireHierarchicalClickHandler
-      }
-    }
-    const args = {
-      points: event.points.map(p =>
-        serializePoint(
-          p as unknown as Record<string, unknown> & {
-            data: Partial<PlotData>;
-            curveNumber: number;
-          }
-        )
-      ),
-      modifiers: this.getModifiers(),
-    };
-    this.sendEventCallback(clickId, args);
-  }
-
-  /**
-   * Handle a box/lasso selection.
-   */
-  onSelected(event: PlotSelectionEvent | undefined): void {
-    const selectedId = this.callbackMap.get('on_selected');
-    if (selectedId == null || event == null) {
-      return;
-    }
-    const args: Record<string, unknown> = {
-      points: event.points.map(p =>
-        serializePoint(
-          p as unknown as Record<string, unknown> & {
-            data: Partial<PlotData>;
-            curveNumber: number;
-          }
-        )
-      ),
-      modifiers: this.getModifiers(),
-    };
-    if (event.range != null) {
-      args.range = event.range;
-    }
-    if (event.lassoPoints != null) {
-      args.lasso_points = event.lassoPoints;
-    }
-    this.sendEventCallback(selectedId, args);
-  }
-
-  /**
-   * Handle a selection being cleared.
-   */
-  onDeselect(): void {
-    const deselectId = this.callbackMap.get('on_deselect');
-    if (deselectId == null) {
-      return;
-    }
-    this.sendEventCallback(deselectId, { modifiers: this.getModifiers() });
-  }
-
-  /**
-   * Handle an annotation click.
-   */
-  onClickAnnotation(event: ClickAnnotationEvent): void {
-    const annotationId = this.callbackMap.get('on_click_annotation');
-    if (annotationId == null) {
-      return;
-    }
-    const args = {
-      index: event.index,
-      annotation: {
-        text: event.annotation.text,
-        x: event.annotation.x,
-        y: event.annotation.y,
-      },
-      modifiers: this.getModifiers(),
-    };
-    this.sendEventCallback(annotationId, args);
-  }
-
-  /**
-   * Timer for debouncing legend single-click vs double-click.
-   * When on_legend_click is registered, single clicks are held for
-   * LEGEND_DBL_CLICK_DELAY ms so a double-click can cancel them.
-   */
-  private legendClickTimer: ReturnType<typeof setTimeout> | null = null;
-
-  /**
-   * Handle a legend item click.
-   * When on_legend_click is registered the click is debounced to
-   * discriminate from double-clicks (Plotly fires plotly_legendclick
-   * for BOTH clicks of a double-click). If no double-click arrives
-   * within LEGEND_DBL_CLICK_DELAY ms the single-click is sent to Python.
-   * Returns false to prevent Plotly's default toggle.
-   */
-  onLegendClick(event: LegendClickEvent): boolean {
-    const legendClickId = this.callbackMap.get('on_legend_click');
-    if (legendClickId == null) {
-      // No callback registered — let Plotly perform its default toggle.
-      return true;
-    }
-    const gd = this.plotElement;
-    const args = {
-      trace_name:
-        (event.data?.[event.curveNumber] as Partial<PlotData>)?.name ?? '',
-      curve_number: event.curveNumber,
-      modifiers: this.getModifiers(),
-    };
-
-    // Cancel any pending debounced click from a previous first-click
-    if (this.legendClickTimer != null) {
-      clearTimeout(this.legendClickTimer);
-      this.legendClickTimer = null;
-    }
-
-    // Debounce: wait to see if a double-click follows
-    this.legendClickTimer = setTimeout(() => {
-      this.legendClickTimer = null;
-      this.sendEventCallbackWithResponse(legendClickId, args).then(allowed => {
-        if (allowed && gd != null) {
-          const currentVis = (gd.data?.[event.curveNumber] as Partial<PlotData>)
-            ?.visible;
-          const nextVis = currentVis === 'legendonly' ? true : 'legendonly';
-          Plotly.restyle(gd, { visible: nextVis }, [event.curveNumber]);
-        }
-      });
-    }, LEGEND_DBL_CLICK_DELAY);
-
-    return false; // Always prevent; re-applied above if allowed.
-  }
-
-  /**
-   * Handle a legend item double-click.
-   * Cancels any pending debounced single-click so only the double-click
-   * fires. Returns false to prevent Plotly's default isolate/show-all,
-   * then re-applies it only if the Python callback allows it.
-   *
-   * When on_legend_click is registered, its handler returns false which also
-   * suppresses Plotly's native double-click, so we reimplement the default here
-   * even when there's no dedicated double-click callback.
-   */
-  onLegendDoubleClick(event: LegendClickEvent): boolean {
-    const legendClickId = this.callbackMap.get('on_legend_click');
-    const legendDblClickId = this.callbackMap.get('on_legend_double_click');
-    if (legendClickId == null && legendDblClickId == null) {
-      // Nothing registered — let Plotly perform its default.
-      return true;
-    }
-
-    // Cancel pending debounced single-click
-    if (this.legendClickTimer != null) {
-      clearTimeout(this.legendClickTimer);
-      this.legendClickTimer = null;
-    }
-
-    const gd = this.plotElement;
-    const { curveNumber } = event;
-    if (legendDblClickId != null) {
-      const args = {
-        trace_name:
-          (event.data?.[curveNumber] as Partial<PlotData>)?.name ?? '',
-        curve_number: curveNumber,
-        modifiers: this.getModifiers(),
-      };
-      this.sendEventCallbackWithResponse(legendDblClickId, args).then(
-        allowed => {
-          if (allowed && gd != null) {
-            PlotlyExpressChartModel.performLegendDoubleClick(gd, curveNumber);
-          }
-        }
-      );
-    } else if (gd != null) {
-      // Only on_legend_click is registered; its false return blocks the native
-      // double-click, so perform the default isolate/show-all directly.
-      PlotlyExpressChartModel.performLegendDoubleClick(gd, curveNumber);
-    }
-    return false; // Always prevent; re-applied above if allowed.
   }
 
   override unsubscribe(callback: (event: ChartEvent) => void): void {
@@ -769,12 +327,6 @@ export class PlotlyExpressChartModel extends ChartModel {
     super.unsubscribe(callback);
     this.widgetUnsubscribe?.();
     this.isSubscribed = false;
-
-    document.removeEventListener(
-      'pointerdown',
-      this.handlePointerModifiers,
-      true
-    );
 
     this.tableReferenceMap.forEach((_, id) => this.removeTable(id));
 
@@ -1495,27 +1047,6 @@ export class PlotlyExpressChartModel extends ChartModel {
     );
   }
 
-  private relayoutTimer: ReturnType<typeof setTimeout> | null = null;
-
-  private relayoutMerged: Record<string, unknown> = {};
-
-  onRelayout(changes: Record<string, unknown>): void {
-    const relayoutId = this.callbackMap.get('on_relayout');
-    if (relayoutId == null) return;
-
-    // Debounce: merge keys from rapid events, send once after 150ms pause
-    Object.assign(this.relayoutMerged, changes);
-    if (this.relayoutTimer) clearTimeout(this.relayoutTimer);
-    this.relayoutTimer = setTimeout(() => {
-      this.sendEventCallback(relayoutId, {
-        ...this.relayoutMerged,
-        modifiers: this.getModifiers(),
-      });
-      this.relayoutMerged = {};
-      this.relayoutTimer = null;
-    }, 150);
-  }
-
   isPreventable(callbackId: string): boolean {
     return this.preventableCallbacks.has(callbackId);
   }
@@ -1561,35 +1092,6 @@ export class PlotlyExpressChartModel extends ChartModel {
     if (resolver) {
       this.pendingResponses.delete(parsed.request_id);
       resolver(parsed.result !== false);
-    }
-  }
-
-  /**
-   * Perform the default legend double-click behavior: toggle between
-   * isolating the clicked trace and showing all.
-   */
-  private static performLegendDoubleClick(
-    gd: PlotlyGraphDiv,
-    curveNumber: number
-  ): void {
-    const traces = gd.data ?? [];
-    const isAlreadyIsolated = traces.every(
-      (trace, i) =>
-        i === curveNumber ||
-        (trace as Partial<PlotData>).visible === false ||
-        (trace as Partial<PlotData>).visible === 'legendonly'
-    );
-
-    if (isAlreadyIsolated) {
-      // Show all
-      Plotly.restyle(gd, { visible: true });
-    } else {
-      // Isolate — hide all others
-      const visibilities = traces.map((trace, i) => {
-        if ((trace as Partial<PlotData>).visible === false) return false; // false is sticky
-        return i === curveNumber ? true : 'legendonly';
-      });
-      Plotly.restyle(gd, { visible: visibilities } as unknown as Data);
     }
   }
 }
